Trigger CircleCI build from Netlify
This is a hack, as CircleCI is triggered during the Netlify Build, and waits 70 seconds before running Cypress tests.
Ideally, we should set up a server (or serverless function)
that would be used as a Netlify webhook notification.
This server (or serverless function) would then call CircleCI
with the right parameters